Move scrolling tests to their own file
